feat(smartview): add `eval`.
Allows to evaluate a java-script expression and capture its result.
The java-script expression is a format-string that can capture the arguments passed as args.
To reduce the amount of code the user needs to write one can also use saucer::arguments as an argument to reduce redundancy.
A call to `Math.pow` for example can thus look like this:
`smartview.eval<int>("Math.pow({},{})", 2, 4);`
or like this:
`smartview.eval<int>("Math.pow({})", saucer::make_arguments(2,4));`