fix(traits): parsing of `std::string_view`
Some users might be inclined to expose functions which take a `std::string_view` as parameter.

While the idea makes sense, it is inherently dangerous as having such an argument will instruct the serializer to read from a given buffer into a `string_view` which will be tied to the buffer life-time and as such might be a problem.
Furthermore, some serializers don't even support reading into a `string_view` (rflpp).

Thus we replace all `std::string_view` occurences in the params with `std::string`. This way the serializer will happily (and safely!) parse the given data and users have the convenience of not having to worry about `string_view` parameters :)