refactor(webview): remove launch policy
This commit removes the built-in thread-pool and `saucer::launch`. With the implementation of Coroutines the need for built-in threading was made obsolete.

In case users need threading, they are encouraged to implement it themselves - executors already take care of the lifetimes.

We still depend on `poolparty` for the `packaged_task` implementation - if a thread-pool is required, users could use this directly.