refactor(window/webview, wkg): move to nested `impl` [1/4] [skip ci]
This is part of a bigger refactoring effort to reduce the amount of boilerplate code (e.g `if (!thread_safe) dispatch(...)`).

While this will significantly reduce the boilerplate code, it will also make the addition of move-semantics easier in the future, which among other things will finally allow for proper error-handling on construction!