refactor(unquoted)!: minor improvements
The value returned by unquoted is later used to construct a string
anyways, so it makes more sense to just store a string here directly. We
also disallow passing unquoted_t as an l-value, as we will later simply
move the string out of it. Furthermore, having unquoted be a function
makes it easier to restrict its usage in the future.